Code clean up. Now pref360controller has a new UI.

+++ Pref360Control/MyDeadZoneViewer.h
@@ -10,7 +10,7 @@
 
 @interface MyDeadZoneViewer : NSView
 
-@property (nonatomic) int val;
+@property (nonatomic) double val;
 @property (nonatomic) BOOL linked;
 
 @end
